Check user for votting current product
